requirements.txt
fastapi==0.115.7
loguru==0.7.3
tortoise-orm==0.24.0
uvicorn==0.34.0
pydantic==2.10.6
aiogram==3.17.0
jinja2==3.1.5
torch==2.4.0
asyncpg==0.30.0
python-dotenv==1.0.1
alembic==1.14.0
pytest==8.3.4
python-jose==3.3.0
python-multipart==0.0.20
openai==1.54.3
aerich==0.2.4